Fix NoMethodError when not using NTLM
Issue #115. This fix makes HTTPI.logger available,
during require of adapters.
